extract types into file
